making sure the file being updated or deleted actually belongs to the logged in user

+++ routes/api.js
@@ -117,21 +117,29 @@ router.post('/api/update', utils.isLoggedIn, function (req, res) {
         if (error || !item) {
             res.json({ error: defaultErrorMessage });
         } else {
-            if (_.isArray(body['category[]'])) {
-                item.tags = body['category[]'];
+
+            if (item.uid.toString() == req.user._id.toString()) {
+
+                if (_.isArray(body['category[]'])) {
+                    item.tags = body['category[]'];
+                } else {
+                    item.tags = [ body['category[]'] ];
+                }
+                
+                item.location = body.location;
+
+                db.get().collection('imagePosts').save(item, function (error, result) {
+                    if (error) {
+                        res.json({ error: defaultErrorMessage });
+                    } else {
+                        res.json({ message: 'Successfully updated image! :D' });
+                    }
+                });
+
             } else {
-                item.tags = [ body['category[]'] ];
+                res.json({ error: defaultErrorMessage });
             }
             
-            item.location = body.location;
-
-            db.get().collection('imagePosts').save(item, function (error, result) {
-                if (error) {
-                    res.json({ error: defaultErrorMessage });
-                } else {
-                    res.json({ message: 'Successfully updated image! :D' });
-                }
-            });
         }
 
     }); 
@@ -141,19 +149,32 @@ router.post('/api/update', utils.isLoggedIn, function (req, res) {
 router.get('/api/remove/:uniqueName', utils.isLoggedIn, function (req, res) {
     var defaultErrorMessage = 'Error removing image! :O';
 
-    db.get().collection('imagePosts').deleteOne({ name: req.params.uniqueName }, function (error, results) {
+    db.get().collection('imagePosts').findOne({ name: req.params.uniqueName }, function (error, item) {
 
-        if (error) {
-            res.json({ error: defaultErrorMessage });
+        if (!error && item && item.uid.toString() == req.user._id.toString()) {
+
+            db.get().collection('imagePosts').deleteOne({ name: req.params.uniqueName }, function (error, results) {
+
+                if (error) {
+                    res.json({ error: defaultErrorMessage });
+                } else {
+
+                    s3.getBucket().deleteObject({ Key: req.params.uniqueName }, function (error, data) {
+                        if (error) { res.json({ error: defaultErrorMessage }); }
+
+                        res.json({ message: 'Successfully deleted image! :D' });
+                    });
+                    
+                }
+
+            }); 
+
         } else {
-            s3.getBucket().deleteObject({ Key: req.params.uniqueName }, function (error, data) {
-                if (error) { res.json({ error: defaultErrorMessage }); }
-
-                res.json({ message: 'Successfully deleted image! :D' });
-            });
+            res.json({ error: defaultErrorMessage });
         }
+    });
 
-    }); 
+    
     
 });
